+++ src/main/java/com/projeto/sistema_escolar/repository/AlternativaRepository.java
@@ -4,6 +4,6 @@ import com.projeto.sistema_escolar.model.Alternativa;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface AlternativaRepository extends JpaRepository<Alternativa, Long> {
-    List<Alternativa> findByQuestaoId(Long questaoId);
+public interface AlternativaRepository extends JpaRepository<Alternativa, Integer> {
+    List<Alternativa> findByQuestaoId(Integer questaoId);
 }
