+++ src/main/java/com/projeto/sistema_escolar/repository/AlunoRepository.java
@@ -4,6 +4,6 @@ import com.projeto.sistema_escolar.model.Aluno;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface AlunoRepository extends JpaRepository<Aluno, Long> {
-    Optional<Aluno> findByUsuarioId(Long usuarioId);
+public interface AlunoRepository extends JpaRepository<Aluno, Integer> {
+    Optional<Aluno> findByUsuarioId(Integer usuarioId);
 }
