+++ src/main/java/com/projeto/sistema_escolar/repository/CoordenadorRepository.java
@@ -4,6 +4,6 @@ import com.projeto.sistema_escolar.model.Coordenador;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface CoordenadorRepository extends JpaRepository<Coordenador, Long> {
-    Optional<Coordenador> findByUsuarioId(Long usuarioId);
+public interface CoordenadorRepository extends JpaRepository<Coordenador, Integer> {
+    Optional<Coordenador> findByUsuarioId(Integer usuarioId);
 }
