+++ src/main/java/com/projeto/sistema_escolar/repository/DisciplinaRepository.java
@@ -1,5 +1,7 @@
 package com.projeto.sistema_escolar.repository;
+
 import com.projeto.sistema_escolar.model.Disciplina;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DisciplinaRepository extends JpaRepository<Disciplina, Long> {}
+public interface DisciplinaRepository extends JpaRepository<Disciplina, Integer> {
+}
