+++ src/main/java/com/projeto/sistema_escolar/repository/EscolaRepository.java
@@ -3,5 +3,5 @@ package com.projeto.sistema_escolar.repository;
 import com.projeto.sistema_escolar.model.Escola;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EscolaRepository extends JpaRepository<Escola, Long> {
+public interface EscolaRepository extends JpaRepository<Escola, Integer> {
 }
