+++ src/main/java/com/projeto/sistema_escolar/repository/PerfilRepository.java
@@ -3,5 +3,5 @@ package com.projeto.sistema_escolar.repository;
 import com.projeto.sistema_escolar.model.Perfil;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PerfilRepository extends JpaRepository<Perfil, Long> {
+public interface PerfilRepository extends JpaRepository<Perfil, Integer> {
 }
