+++ src/main/java/com/projeto/sistema_escolar/repository/ProfessorRepository.java
@@ -4,6 +4,6 @@ import com.projeto.sistema_escolar.model.Professor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface ProfessorRepository extends JpaRepository<Professor, Long> {
-    Optional<Professor> findByUsuarioId(Long usuarioId);
+public interface ProfessorRepository extends JpaRepository<Professor, Integer> {
+    Optional<Professor> findByUsuarioId(Integer usuarioId);
 }
