+++ src/main/java/com/projeto/sistema_escolar/repository/ProvaRepository.java
@@ -4,6 +4,6 @@ import com.projeto.sistema_escolar.model.Prova;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface ProvaRepository extends JpaRepository<Prova, Long> {
-    List<Prova> findByProfessorId(Long professorId);
+public interface ProvaRepository extends JpaRepository<Prova, Integer> {
+    List<Prova> findByProfessorId(Integer professorId);
 }
