+++ src/main/java/com/projeto/sistema_escolar/repository/RespostaRepository.java
@@ -4,7 +4,7 @@ import com.projeto.sistema_escolar.model.Resposta;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface RespostaRepository extends JpaRepository<Resposta, Long> {
-    List<Resposta> findByAlunoId(Long alunoId);
-    List<Resposta> findByProvaId(Long provaId);
+public interface RespostaRepository extends JpaRepository<Resposta, Integer> {
+    List<Resposta> findByAlunoId(Integer alunoId);
+    List<Resposta> findByProvaId(Integer provaId);
 }
