+++ src/main/java/com/projeto/sistema_escolar/repository/SerieRepository.java
@@ -1,5 +1,7 @@
 package com.projeto.sistema_escolar.repository;
+
 import com.projeto.sistema_escolar.model.Serie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SerieRepository extends JpaRepository<Serie, Long> {}
+public interface SerieRepository extends JpaRepository<Serie, Integer> {
+}
