+++ src/main/java/com/projeto/sistema_escolar/repository/TokenSessaoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface TokenSessaoRepository extends JpaRepository<TokenSessao, Long> {
+public interface TokenSessaoRepository extends JpaRepository<TokenSessao, Integer> {
     Optional<TokenSessao> findByToken(String token);
-    List<TokenSessao> findByUsuarioId(Long usuarioId);
+    List<TokenSessao> findByUsuarioId(Integer usuarioId);
 }
