+++ src/main/java/com/projeto/sistema_escolar/service/CoordenadorService.java
@@ -3,6 +3,7 @@ package com.projeto.sistema_escolar.service;
 import com.projeto.sistema_escolar.model.Coordenador;
 import com.projeto.sistema_escolar.repository.CoordenadorRepository;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -19,11 +20,11 @@ public class CoordenadorService {
         return repository.findAll();
     }
 
-    public Optional<Coordenador> buscarPorId(Long id) {
+    public Optional<Coordenador> buscarPorId(Integer id) {
         return repository.findById(id);
     }
 
-    public Optional<Coordenador> buscarPorUsuarioId(Long usuarioId) {
+    public Optional<Coordenador> buscarPorUsuarioId(Integer usuarioId) {
         return repository.findByUsuarioId(usuarioId);
     }
 
@@ -31,11 +32,11 @@ public class CoordenadorService {
         return repository.save(coordenador);
     }
 
-    public void deletar(Long id) {
+    public void deletar(Integer id) {
         repository.deleteById(id);
     }
 
-    public boolean existePorId(Long id) {
+    public boolean existePorId(Integer id) {
         return repository.existsById(id);
     }
 }
