+++ src/main/java/com/projeto/sistema_escolar/service/RespostaService.java
@@ -3,6 +3,7 @@ package com.projeto.sistema_escolar.service;
 import com.projeto.sistema_escolar.model.Resposta;
 import com.projeto.sistema_escolar.repository.RespostaRepository;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -19,15 +20,15 @@ public class RespostaService {
         return repository.findAll();
     }
 
-    public Optional<Resposta> buscarPorId(Long id) {
+    public Optional<Resposta> buscarPorId(Integer id) {
         return repository.findById(id);
     }
 
-    public List<Resposta> buscarPorAluno(Long alunoId) {
+    public List<Resposta> buscarPorAluno(Integer alunoId) {
         return repository.findByAlunoId(alunoId);
     }
 
-    public List<Resposta> buscarPorProva(Long provaId) {
+    public List<Resposta> buscarPorProva(Integer provaId) {
         return repository.findByProvaId(provaId);
     }
 
@@ -35,11 +36,11 @@ public class RespostaService {
         return repository.save(resposta);
     }
 
-    public void deletar(Long id) {
+    public void deletar(Integer id) {
         repository.deleteById(id);
     }
 
-    public boolean existePorId(Long id) {
+    public boolean existePorId(Integer id) {
         return repository.existsById(id);
     }
 }
