+++ src/main/java/com/projeto/sistema_escolar/service/SerieService.java
@@ -3,6 +3,7 @@ package com.projeto.sistema_escolar.service;
 import com.projeto.sistema_escolar.model.Serie;
 import com.projeto.sistema_escolar.repository.SerieRepository;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -19,7 +20,7 @@ public class SerieService {
         return repository.findAll();
     }
 
-    public Optional<Serie> buscarPorId(Long id) {
+    public Optional<Serie> buscarPorId(Integer id) {
         return repository.findById(id);
     }
 
@@ -27,11 +28,11 @@ public class SerieService {
         return repository.save(serie);
     }
 
-    public void deletar(Long id) {
+    public void deletar(Integer id) {
         repository.deleteById(id);
     }
 
-    public boolean existePorId(Long id) {
+    public boolean existePorId(Integer id) {
         return repository.existsById(id);
     }
 }
